✨  🛠️ refactor: pr 리뷰에 맞춰 수정함 ver1
🔗 Resolves: #3
