+++ src/main/java/com/kbe5/rento/domain/device/service/DeviceEventService.java
@@ -11,8 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service
-@Transactional(readOnly = true)
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class DeviceEventService {
 
     private final DeviceEventRepository deviceEventRepository;
