+++ src/main/java/com/kbe5/rento/domain/device/service/DeviceService.java
@@ -10,8 +10,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service
-@Transactional(readOnly = true)
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class DeviceService {
 
     private final DeviceRepository deviceRepository;
